Rearrange integration tests into a clearer folder hierarchy: not test cases have been added or changed
