+++ integration_tests/pages/reports/involvements/abstract/index.ts
@@ -1,0 +1,4 @@
+export * from './add'
+export * from './involvements'
+export * from './remove'
+export * from './search'
