+++ integration_tests/pages/reports/involvements/abstract/remove.ts
@@ -1,0 +1,8 @@
+import FormWizardPage from '../../../formWizard'
+
+// eslint-disable-next-line import/prefer-default-export
+export abstract class RemoveInvolvementsPage extends FormWizardPage {
+  selectRadioButton(text: string) {
+    this.radioOrCheckboxButton('confirmRemove', text).click()
+  }
+}
