+++ integration_tests/pages/reports/involvements/staff/index.ts
@@ -1,0 +1,5 @@
+export * from './add'
+export * from './involvements'
+export * from './manual'
+export * from './remove'
+export * from './search'
