+++ integration_tests/pages/reports/involvements/staff/remove.ts
@@ -1,0 +1,8 @@
+import { RemoveInvolvementsPage } from '../abstract'
+
+// eslint-disable-next-line import/prefer-default-export
+export class RemoveStaffInvolvementsPage extends RemoveInvolvementsPage {
+  constructor(name: string) {
+    super(`Are you sure you want to remove ${name}?`, 'Remove member of staff from report')
+  }
+}
