+++ integration_tests/pages/reports/involvements/staff/search.ts
@@ -1,6 +1,7 @@
-import FormWizardPage from '../../formWizard'
+import { SearchPage } from '../abstract'
 
-export default class StaffSearchPage extends FormWizardPage {
+// eslint-disable-next-line import/prefer-default-export
+export class StaffSearchPage extends SearchPage {
   constructor(results?: { found: true } | { notFound: string }) {
     let pageTitle: string
     if (!results) {
@@ -25,10 +26,6 @@ export default class StaffSearchPage extends FormWizardPage {
     return cy.get<HTMLAnchorElement>('a').contains('manually add the member of staff to the report').end()
   }
 
-  showsNoTable(): Cypress.Chainable<void> {
-    return cy.get<HTMLTableRowElement>('.app-involvement-search-results').should('not.exist').end()
-  }
-
   get tableContents(): Cypress.Chainable<
     {
       name: string
@@ -38,7 +35,7 @@ export default class StaffSearchPage extends FormWizardPage {
       actionLink: HTMLAnchorElement
     }[]
   > {
-    return cy.get<HTMLTableRowElement>('table.app-involvement-search-results tbody tr.govuk-table__row').then(rows =>
+    return this.tableRows.then(rows =>
       rows
         .map((_, row) => {
           const $cells = Cypress.$(row).find('.govuk-table__cell') as unknown as JQuery<HTMLTableCellElement>
